Split Wilkinsburg borough and school district EIT, fix calculation
The calculation for "school district merger only" was double counting
school taxes. This splits out the 1% Wilkinsburg EIT into its borough
and school district components, and updates the calculations.

The calculation for "no school district merger" was undercounting but
not including the Wilkinsburg school district EIT that would remain in
place.

This also re-groups the calculations to make the code easier to
follow. I can un-do this if you'd prefer; for me it was necessary to
understand the short variable names.