Breaks out Wilkinsburg muni and SD EIT
I had to abandon the GPVAL_Z_MIN/MAX stuff because it didn't obey my
zero point in the palette spec. It's more work to maintain the hardcoded
figures but I'll incur that cost in exchange for correctness. After this
change, the taxation amounts _should_ be accurate and only change if
millage changes...

Fixes #5